backend-api/requirements.txt
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1

# LangChain for semantic search
langchain==0.3.7
langchain-openai==0.2.6
langchain-community==0.3.7

# Vector database options
chromadb==0.5.18
faiss-cpu==1.9.0

# For embeddings
tiktoken==0.8.0

